Extracted db queries into separate files, fixed a bug where profiles couldn't be added.
